Push mit syntaxfehler

+++ src/main/java/de/supercode/friends/repository/GroupRepo.java
@@ -3,5 +3,5 @@ package de.supercode.friends.repository;
 import de.supercode.friends.entities.Group;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GroupRepo extends JpaRepository<Group,Long> {
+public interface GroupRepo extends JpaRepository<Group,Integer> {
 }
